Added GenericSnackbar and error for no card
