+++ autoscheduler/frontend/src/components/SchedulingPage/Schedule/MeetingCard/MeetingCard.tsx
@@ -7,6 +7,7 @@ import Meeting, { MeetingType } from '../../../../types/Meeting';
 import ScheduleCard from '../ScheduleCard/ScheduleCard';
 import { generateSectionInfoID } from '../../CourseSelectColumn/CourseSelectCard/ExpandedCourseCard/SectionSelect/SectionInfo';
 import { scrollSectionIntoView } from '../../../../redux/actions/courseCards';
+import GenericSnackbar from '../../../GenericSnackbar';
 
 interface MeetingCardProps {
   meeting: Meeting;
@@ -27,6 +28,7 @@ const MeetingCard: React.FC<MeetingCardProps> = ({
     startTimeHours, startTimeMinutes, endTimeHours, endTimeMinutes, section, meetingType,
   } = meeting;
   const [isBig, setIsBig] = React.useState(true);
+  const [snackbarMessage, setSnackbarMessage] = React.useState('');
   const dispatch = useDispatch();
 
   // hide meeting type if the card is small
@@ -50,7 +52,12 @@ const MeetingCard: React.FC<MeetingCardProps> = ({
       }, 1000);
     }
 
-    dispatch(scrollSectionIntoView(meeting.section));
+    try {
+      dispatch(scrollSectionIntoView(meeting.section));
+    } catch (error) {
+      setSnackbarMessage(error.message);
+      return;
+    }
 
     setTimeout(() => {
       executeScroll();
@@ -58,35 +65,38 @@ const MeetingCard: React.FC<MeetingCardProps> = ({
   };
 
   return (
-    <ScheduleCard
-      startTimeHours={startTimeHours}
-      startTimeMinutes={startTimeMinutes}
-      endTimeHours={endTimeHours}
-      endTimeMinutes={endTimeMinutes}
-      firstHour={firstHour}
-      lastHour={lastHour}
-      onResizeWindow={
-        (contentHeight, clientHeight): void => handleResize(contentHeight < clientHeight)
-      }
-      backgroundColor={bgColor}
-      borderColor={bgColor}
-      onClick={onClick}
-    >
-      <Typography variant="body2" data-testid="meeting-card-primary-content" className={styles.meetingCardText}>
-        {`${section.subject} ${section.courseNum}`}
-        {isBig
-          ? `-${section.sectionNum}`
-          : (
-            <Typography variant="subtitle2" component="span">
-              &nbsp;
-              {`${MeetingType[meetingType]}`}
-            </Typography>
-          )}
-      </Typography>
-      <Typography variant="subtitle2" style={{ display: isBig ? 'block' : 'none' }}>
-        {MeetingType[meetingType]}
-      </Typography>
-    </ScheduleCard>
+    <>
+      <ScheduleCard
+        startTimeHours={startTimeHours}
+        startTimeMinutes={startTimeMinutes}
+        endTimeHours={endTimeHours}
+        endTimeMinutes={endTimeMinutes}
+        firstHour={firstHour}
+        lastHour={lastHour}
+        onResizeWindow={
+          (contentHeight, clientHeight): void => handleResize(contentHeight < clientHeight)
+        }
+        backgroundColor={bgColor}
+        borderColor={bgColor}
+        onClick={onClick}
+      >
+        <Typography variant="body2" data-testid="meeting-card-primary-content" className={styles.meetingCardText}>
+          {`${section.subject} ${section.courseNum}`}
+          {isBig
+            ? `-${section.sectionNum}`
+            : (
+              <Typography variant="subtitle2" component="span">
+                &nbsp;
+                {`${MeetingType[meetingType]}`}
+              </Typography>
+            )}
+        </Typography>
+        <Typography variant="subtitle2" style={{ display: isBig ? 'block' : 'none' }}>
+          {MeetingType[meetingType]}
+        </Typography>
+      </ScheduleCard>
+      <GenericSnackbar snackbarMessage={snackbarMessage} setSnackbarMessage={setSnackbarMessage} />
+    </>
   );
 };
